@@ -2,6 +2,5 @@ package com.problem_solving.drawing.utils.constants;
 
 public enum Type {
     
-    Canvas, Line, Rectangle, BucketFilling
-    
+    Canvas, Line, Rectangle, BucketFilling    
 }
