@@ -11,7 +11,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Unexpected number of command line args");
 
-        String commandLine = "R 16 1 20 3 5";
+        var commandLine = "R 16 1 20 3 5";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -21,7 +21,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Unexpected number of command line args");
 
-        String commandLine = "R 16 1 20";
+        var commandLine = "R 16 1 20";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -32,7 +32,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 1 whose value is X");
 
-        String commandLine = "R X 1 20 3";
+        var commandLine = "R X 1 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -42,7 +42,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 2 whose value is X");
 
-        String commandLine = "R 16 X 20 3";
+        var commandLine = "R 16 X 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -52,7 +52,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 3 whose value is X");
 
-        String commandLine = "R 16 1 X 3";
+        var commandLine = "R 16 1 X 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -62,7 +62,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 4 whose value is X");
 
-        String commandLine = "R 16 1 20 X";
+        var commandLine = "R 16 1 20 X";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -73,7 +73,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 1 whose value is -16");
 
-        String commandLine = "R -16 1 20 3";
+        var commandLine = "R -16 1 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -83,7 +83,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 2 whose value is -1");
 
-        String commandLine = "R 16 -1 20 3";
+        var commandLine = "R 16 -1 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -93,7 +93,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 3 whose value is -20");
 
-        String commandLine = "R 16 1 -20 3";
+        var commandLine = "R 16 1 -20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -103,7 +103,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 4 whose value is -3");
 
-        String commandLine = "R 16 1 20 -3";
+        var commandLine = "R 16 1 20 -3";
         commandValidator.validateCommand(commandLine);
     }
     
@@ -114,7 +114,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 1 whose value is 0 should be greater than zero");
 
-        String commandLine = "R 0 1 20 3";
+        var commandLine = "R 0 1 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -124,7 +124,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 2 whose value is 0 should be greater than zero");
 
-        String commandLine = "R 16 0 20 3";
+        var commandLine = "R 16 0 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -134,7 +134,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 3 whose value is 0 should be greater than zero");
 
-        String commandLine = "R 16 1 0 3";
+        var commandLine = "R 16 1 0 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -144,7 +144,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 4 whose value is 0 should be greater than zero");
 
-        String commandLine = "R 16 1 20 0";
+        var commandLine = "R 16 1 20 0";
         commandValidator.validateCommand(commandLine);
     }
     
@@ -155,7 +155,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 1 whose value is 16.5");
 
-        String commandLine = "R 16.5 1 20 3";
+        var commandLine = "R 16.5 1 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -165,7 +165,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 2 whose value is 1.5");
 
-        String commandLine = "R 16 1.5 20 3";
+        var commandLine = "R 16 1.5 20 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -175,7 +175,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 3 whose value is 20.5");
 
-        String commandLine = "R 16 1 20.5 3";
+        var commandLine = "R 16 1 20.5 3";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -185,7 +185,7 @@ public class RectangleCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 4 whose value is 3.5");
 
-        String commandLine = "R 16 1 20 3.5";
+        var commandLine = "R 16 1 20 3.5";
         commandValidator.validateCommand(commandLine);
     }
     
