@@ -11,7 +11,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Unexpected number of command line args");
 
-        String commandLine = "L 6 3 6 4 8";
+        var commandLine = "L 6 3 6 4 8";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -21,7 +21,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Unexpected number of command line args");
 
-        String commandLine = "L 6 3 6";
+        var commandLine = "L 6 3 6";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -32,7 +32,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 1 whose value is X");
 
-        String commandLine = "L X 3 6 4";
+        var commandLine = "L X 3 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -42,7 +42,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 2 whose value is X");
 
-        String commandLine = "L 6 X 6 4";
+        var commandLine = "L 6 X 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -52,7 +52,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 3 whose value is X");
 
-        String commandLine = "L 6 3 X 4";
+        var commandLine = "L 6 3 X 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -62,7 +62,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 4 whose value is X");
 
-        String commandLine = "L 6 3 6 X";
+        var commandLine = "L 6 3 6 X";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -73,7 +73,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 1 whose value is -6");
 
-        String commandLine = "L -6 3 6 4";
+        var commandLine = "L -6 3 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -83,7 +83,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 2 whose value is -3");
 
-        String commandLine = "L 6 -3 6 4";
+        var commandLine = "L 6 -3 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -93,7 +93,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 3 whose value is -6");
 
-        String commandLine = "L 6 3 -6 4";
+        var commandLine = "L 6 3 -6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -103,7 +103,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 4 whose value is -4");
 
-        String commandLine = "L 6 3 6 -4";
+        var commandLine = "L 6 3 6 -4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -114,7 +114,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 1 whose value is 0 should be greater than zero");
 
-        String commandLine = "L 0 3 6 4";
+        var commandLine = "L 0 3 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -124,7 +124,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 2 whose value is 0 should be greater than zero");
 
-        String commandLine = "L 6 0 6 4";
+        var commandLine = "L 6 0 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -134,7 +134,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 3 whose value is 0 should be greater than zero");
 
-        String commandLine = "L 6 3 0 4";
+        var commandLine = "L 6 3 0 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -144,7 +144,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Value for argument number 4 whose value is 0 should be greater than zero");
 
-        String commandLine = "L 6 3 6 0";
+        var commandLine = "L 6 3 6 0";
         commandValidator.validateCommand(commandLine);
     }
     
@@ -155,7 +155,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 1 whose value is 1.5");
 
-        String commandLine = "L 1.5 3 6 4";
+        var commandLine = "L 1.5 3 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -165,7 +165,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 2 whose value is 2.5");
 
-        String commandLine = "L 6 2.5 6 4";
+        var commandLine = "L 6 2.5 6 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -175,7 +175,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 3 whose value is 6.5");
 
-        String commandLine = "L 6 3 6.5 4";
+        var commandLine = "L 6 3 6.5 4";
         commandValidator.validateCommand(commandLine);
     }
 
@@ -185,7 +185,7 @@ public class VerticalLineCommandTest extends BaseCommandValidatorTest {
         thrownException.expect(IllegalArgumentException.class);
         thrownException.expectMessage("Invalid value for argument number 4 whose value is 4.5");
 
-        String commandLine = "L 6 3 6 4.5";
+        var commandLine = "L 6 3 6 4.5";
         commandValidator.validateCommand(commandLine);
     }
     
