@@ -18,9 +18,9 @@ public class VanyaAndFenceProblemTest {
     public void testFindTheMinimumPossibleValidWidthOfTheRoad() {
         
         int n = 6, h = 5;
-        int[] heights = {7, 6, 8, 9, 10, 5};
+        var heights = new int[]{7, 6, 8, 9, 10, 5};
         
-        int result = vafp.findTheMinimumPossibleValidWidthOfTheRoad(n, h, heights);
+        var result = vafp.findTheMinimumPossibleValidWidthOfTheRoad(n, h, heights);
         
         assertEquals(11, result);        
     }
